CarParkingSystem
It enables vehicles to detect objects around them while parking and to warn the driver. The parking sensor measures the distance between the vehicle and obstacles during parking maneuvers thanks to the sensors placed in front and/or behind the vehicles. In this way, drivers can park more safely and easily.